Bugfix - replace `is` with `==`
The python `is` keyword is used to refer to the same object in memory
whereas `==` is used to compare values. In this case, since we're concerned
with values, this should be used.

This commit also adds the ignore for F841 to ensure ruff doesn't complain
about the `dont_display` not being used. This can potentially be a
default value, but that would require adding in a db migration, which can
be considered later.